# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CircleCue API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/me": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get Me",
        "operationId": "get_me_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Delete Account",
        "operationId": "delete_account_me_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "users"
        ],
        "summary": "Update Me",
        "operationId": "update_me_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/code-lookup": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Lookup Code",
        "operationId": "lookup_code_users_code_lookup_get",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodeLookupResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me/routine": {
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Set Routine",
        "operationId": "set_routine_me_routine_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me/pause": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Pause Sharing",
        "operationId": "pause_sharing_me_pause_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/PauseRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Req"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Unpause Sharing",
        "operationId": "unpause_sharing_me_pause_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AIPrefs": {
        "properties": {
          "store_raw": {
            "type": "boolean",
            "title": "Store Raw",
            "default": false
          }
        },
        "type": "object",
        "title": "AIPrefs"
      },
...
      "CodeLookupResponse": {
        "properties": {
          "user_code": {
            "type": "string",
            "title": "User Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          }
        },
        "type": "object",
        "required": [
          "user_code",
          "name"
        ],
        "title": "CodeLookupResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PauseRequest": {
        "properties": {
          "until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Until"
          }
        },
        "type": "object",
        "title": "PauseRequest"
      },
...
      "RoutinePrefs": {
        "properties": {
          "wake": {
            "type": "string",
            "title": "Wake",
            "default": "07:00"
          },
          "sleep": {
            "type": "string",
            "title": "Sleep",
            "default": "23:00"
          },
          "min_call_window_min": {
            "type": "integer",
            "title": "Min Call Window Min",
            "default": 10
          },
          "buffer_min": {
            "type": "integer",
            "title": "Buffer Min",
            "default": 5
          }
        },
        "type": "object",
        "title": "RoutinePrefs"
      },
      "RoutineRequest": {
        "properties": {
          "wake": {
            "type": "string",
            "title": "Wake",
            "default": "07:00"
          },
          "sleep": {
            "type": "string",
            "title": "Sleep",
            "default": "23:00"
          },
          "min_call_window_min": {
            "type": "integer",
            "title": "Min Call Window Min",
            "default": 10
          },
          "buffer_min": {
            "type": "integer",
            "title": "Buffer Min",
            "default": 5
          }
        },
        "type": "object",
        "title": "RoutineRequest"
      },
...
      "SharingPaused": {
        "properties": {
          "active": {
            "type": "boolean",
            "title": "Active",
            "default": false
          },
          "until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Until"
          }
        },
        "type": "object",
        "title": "SharingPaused"
      },
...
      "UpdateProfileRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "tz": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Tz"
          },
          "routine_prefs": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RoutinePrefs"
              },
              {
                "type": "null"
              }
            ]
          },
          "ai_prefs": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AIPrefs"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "UpdateProfileRequest"
      },
      "UserProfileResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "user_code": {
            "type": "string",
            "title": "User Code"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "tz": {
            "type": "string",
            "title": "Tz"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "routine_prefs": {
            "$ref": "#/components/schemas/RoutinePrefs"
          },
          "ai_prefs": {
            "$ref": "#/components/schemas/AIPrefs"
          },
          "sharing_paused": {
            "$ref": "#/components/schemas/SharingPaused"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "user_code",
          "email",
          "tz",
          "routine_prefs",
          "ai_prefs",
          "sharing_paused",
          "created_at"
        ],
        "title": "UserProfileResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}